refactor(dct): Refactor DCT to better handle SIMD lanes
This commit refactors the discrete cosine transform (DCT) implementation to better handle SIMD lanes. The previous implementation iterated over the columns, which was inefficient. This change introduces a `full` const generic parameter to the `do_dct` and `do_idct` functions, which allows the compiler to generate more efficient code for the common case where the number of columns is a multiple of the SIMD vector length.