Unify the JxlOutputBuffer type with the Image type. (#409)
They now belong to the same hierarchy of Image types.

This introduces new RawImage types, that store data independently of the
actual data type used. This will be useful as a replacement for
RowBuffer and for storing pipeline input without needing to generate
code specific to every data type.

The underlying implementation is somewhat more efficient, giving an
approximate 5% speed increase.