Fix max property for references (#438)
The max_property() returns the maximum property index.
The number of property is max_property() + 1 for the property_buffer.

I renamed the function to max_property_count to be more explicit.

Do you want me to add an simple image that triggers this error in the tests ?
I have never created a jxl image but I can take a look into it.
Do you have any recommendation if so ?

Fix #436
Fix #429