Add simple fuzzers via cargo-fuzz (decode, decode_header) (#527)
* Add simple fuzzers via cargo-fuzz (decode, decode_header)

This setup is very primitive (-> not aware of the JPEG-XL file format)
but still finds a bunch of issues quickly:

- unreachable!() on JxlColorEncoding:::XYB in non_section.rs
- assert in huffman.rs: al_size < (1 << HUFFMAN_MAX_BITS)
- debug asserts (bit_reader.rs num <= MAX_BITS_PER_CALL)
- debug panic on integer overflow in huffman.rs:466
(HuffmanCodes::decode decode_varint16(br)? + 1)
- memory limit after allocating an unchecked size (u32 num_toc_entries)
- timeout in infinite / non-progressing loop

Note: The fuzzer is not effective if the target is very crashy, so I
applied workarounds for the panics to get to the list of issues above.

* CONTRIBUTING.md: fix typo

* fuzzer: add issues as failing tests